+++ charts/fluid/fluid/CHANGELOG.md
@@ -18,3 +18,4 @@
 
 * Add debug info for csi
 * Make mount root configurable
+* Update HCFS URL